source/group_eq_bot/requirements.txt
python-telegram-bot==20.0a2
pydantic~=1.9.1
hydra-core==1.2.0
omegaconf~=2.2.3
elasticsearch-dsl>=7.0.0,<8.0.0
shortuuid==1.0.10
